+++ src/pages/Store.jsx
@@ -1,20 +1,10 @@
 import React, { useEffect, useState } from "react";
 import useProductStore from "@/stores/productStore";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import ProductCard from "@/components/product/ProductCard"; 
 import { useNavigate } from "react-router-dom";
-import { ShoppingCart, Heart } from "lucide-react";
-import useCartStore from "@/stores/cartStore";
 
 const Store = () => {
   const { products, actionGetAllProducts } = useProductStore();
-  const addToCart = useCartStore((state) => state.addToCart); // Use addToCart from useCartStore
   const [filteredProducts, setFilteredProducts] = useState([]);
   const [category, setCategory] = useState([]);
   const [categories, setCategories] = useState([]);
@@ -65,23 +55,6 @@ const Store = () => {
     setFilteredProducts(sorted);
   };
 
-  const truncateText = (text, maxLength) => {
-    if (text.length > maxLength) {
-      return text.slice(0, maxLength) + "...";
-    }
-    return text;
-  };
-
-  // Function to handle card click
-  const handleCardClick = (productId) => {
-    navigate(`/product/${productId}`);
-  };
-
-  // Adding product to the cart
-  const handleAddToCart = (product) => {
-    addToCart({ ...product, quantity: 1 });
-  };
-
   return (
     <div className="flex flex-col sm:flex-row gap-6 pt-8 px-4 max-w-[1440px] mx-auto">
       {/* Filter Sidebar */}
@@ -120,46 +93,7 @@ const Store = () => {
         </div>
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
           {filteredProducts.map((product) => (
-            <Card
-              key={product.id}
-              className="border-none shadow-none overflow-hidden rounded-md w-[272px] p-4"
-            >
-              <CardHeader className="p-0">
-                {product.ProductImages && product.ProductImages.length > 0 ? (
-                  <img
-                    src={product.ProductImages[0].imageUrl}
-                    alt={product.name}
-                    className="w-[240px] h-[240px] object-cover rounded-md"
-                    onClick={() => handleCardClick(product.id)}
-                  />
-                ) : (
-                  <img
-                    src="https://via.placeholder.com/150"
-                    alt="No Image Available"
-                    className="w-[240px] h-[240px] object-cover rounded-md"
-                    onClick={() => handleCardClick(product.id)}
-                  />
-                )}
-              </CardHeader>
-              <CardContent className="h-26 p-0 mt-2">
-                <CardDescription>{product.ProductCategory?.name}</CardDescription>
-                <CardTitle className="mt-2 text-lg">
-                  {truncateText(product.name, 24)}
-                </CardTitle>
-                <p className="py-1">{truncateText(product.description, 46)}</p>
-              </CardContent>
-              <CardFooter className="text-lg font-bold p-0 py-2 flex flex-row items-center justify-between">
-                <div className="text-lg">THB {product.price}</div>
-                <div className="flex space-x-2">
-                  <button onClick={() => handleAddToCart(product)}>
-                    <ShoppingCart className="w-6 h-6 hover:scale-110 transition-transform hover:text-blue-500" />
-                  </button>
-                  <button>
-                    <Heart className="w-6 h-6 hover:scale-110 transition-transform hover:text-red-500" />
-                  </button>
-                </div>
-              </CardFooter>
-            </Card>
+            <ProductCard key={product.id} product={product} /> 
           ))}
         </div>
       </div>
